app/requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
Flask-cors
python-dotenv
Flask-JWT-Extended
PyJWT